+++ CHANGELOG.md
@@ -15,6 +15,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Add `ArrayBytesOptional`
 - Add `DataType::Optional`
 - Add `DataType::optional()`
+- Add `FillValue::new_optional_none()`
+- Add `FillValue::optional()`
 - Implement `Element` for `Option<T>` where `T: Element`
 - Implement `ElementOwned` for `Option<T>` where `T: ElementOwned`
 - Implement `ElementFixedLength` for `Option<T>` where `T: ElementFixedLength`